Reorganized test suites to have clearer naming
