+++ backend_plugin/src/test/java/evaluators/AddImportsEvaluatorTest.java
@@ -1,4 +1,4 @@
-package test.java;
+package test.java.evaluators;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
