+++ backend_plugin/src/test/java/evaluators/positiveCases/RemoveImportPositiveCases.java
@@ -1,6 +1,5 @@
-package test.java;
+package test.java.evaluators.positiveCases;
 
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import org.eclipse.jface.text.Position;
@@ -11,7 +10,7 @@ import org.junit.Test;
 
 import main.evaluators.RemoveImportEvaluator;
 
-public class RemoveImportEvaluatorTest {
+public class RemoveImportPositiveCases {
 
 	/** Mock unused import annotation */
 	private static final Annotation ANNOTATION = new Annotation("org.eclipse.jdt.ui.warning",
@@ -42,14 +41,5 @@ public class RemoveImportEvaluatorTest {
 		am.addAnnotation(ANNOTATION, POSITION);
 		eval.evaluateAnnotationModelChanges(am);
 		assertTrue(eval.hasActiveUnusedImportStatement());
-	}
-
-	/**
-	 * Tests that a document being saved in the worspace while this document does not have any
-	 * unused imports does not trigger the feature evaluation
-	 */
-	@Test
-	public void saveEventWithoutUnusedImports() {
-		assertFalse(eval.hasActiveUnusedImportStatement());
 	}
 }
